fix(signer): allow self-issued certificates
According to RFC 5280, Section 3.2, "Self-issued certificates are CA
certificates in which the issuer and subject are the same entity.
Self-issued certificates are generated to support changes in policy or
operations."

Section 6.1 states: "A certificate is self-issued if the same DN
appears in the subject and issuer fields (the two DNs are the same if
they match according to the rules specified in Section 7.1).
In general, the issuer and subject of the certificates that make up a
path are different for each certificate. However, a CA may issue a
certificate to itself to support key rollover or changes in certificate
policies. These self-issued certificates are not counted when
evaluating path length or name constraints."

Section 4.2.1.9 reads: "The pathLenConstraint field is meaningful only
if the cA boolean is asserted and the key usage extension, if present,
asserts the keyCertSign bit (Section 4.2.1.3).  In this case, it gives
the maximum number of non-self-issued intermediate certificates that
may follow this certificate in a valid certification path."

Thus, it should be possible to issue a CA certificate to oneself even if
the pathLenConstraint is zero and clients should be able to construct a
valid certification path.